+++ src/main/java/com/umc/smupool/global/config/Security/PrincipalDetailsService.java
@@ -1,11 +1,10 @@
-package com.umc.smupool.global.config;
+package com.umc.smupool.global.config.Security;
 
 import com.umc.smupool.domain.member.entity.Member;
 import com.umc.smupool.domain.member.exception.MemberErrorCode;
 import com.umc.smupool.domain.member.exception.handler.MemberHandler;
 import com.umc.smupool.domain.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
